+++ packages/core-react/src/index.ts
@@ -1,14 +1,4 @@
 export * from './walletData';
 export * from './hooks';
-
-export type { AcceptTermsModalContextValue } from './context/acceptTermsModal.js';
-export {
-  AcceptTermsModalContext,
-  AcceptTermsModalContextProvider,
-} from './context/acceptTermsModal.js';
-
-export type {
-  ReefKnotContextValue,
-  ReefKnotContextProps,
-} from './context/reefKnot';
-export { ReefKnotContext, ReefKnot } from './context/reefKnot';
+export * from './context';
+export * from './constants';