+++ packages/core-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/core-react
 
+## 1.5.1
+
+### Patch Changes
+
+- Move TERMS_ACCEPTANCE_LS_K const into core-react package
+
 ## 1.5.0
 
 ### Minor Changes