+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,15 @@
 # reef-knot
 
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @reef-knot/web3-react@1.4.1
+  - @reef-knot/connect-wallet-modal@1.6.1
+  - @reef-knot/core-react@1.5.1
+
 ## 1.6.2
 
 ### Patch Changes